Move TwigContext::fromRequest into Twig
Also remove that class.
